solve duplicate insertations
solve duplicate insert by checking database

+++ .NET Core Project/StoreAPICore/Controller/CustomerController.cs
@@ -16,7 +16,7 @@ namespace StoreAPICore.Controller
         string CreateCustomer(InputCustomer customer);
         string UpdateCustomer(Customer customer);
         string DeleteCustomer(Guid customerId);
-        List<Customer> GetAllCustomers();
+        DataTable GetAllCustomers();
 
     }
 
@@ -27,31 +27,63 @@ namespace StoreAPICore.Controller
         string message;
         public string CreateCustomer(InputCustomer customer)
         {
-            
+            List<Customer> customerList = new List<Customer>();
+
             try
             {
                 conn = new DBConnection();
+
+                //Check Customer Already created by mail address
+                customerList = customerDAO.GetAllCustomers(conn);
+
+                bool flagMail = false;
+                bool flagUserName = false;
+
+                //Check User Name Already Exists 
+                foreach (Customer cus in customerList) 
+                {
+                    if (cus.Email == customer.Email)
+                    {
+                        flagMail = true;
+                    }
+
+                    if (cus.UserName == customer.UserName) 
+                    {
+                        flagUserName = true;
+                    }
+                }
+
+                if (flagMail) 
+                {
+                    message = "Mail Address is already Exists..";
+                }
+                else if (flagUserName) 
+                {
+                    message = "User name is already Exists..";
+                }
+                else
+                {
+                    Guid cusId = Guid.NewGuid();
+
+                    Customer user = new Customer();
+                    user.UserId = cusId;
+                    user.UserName = customer.UserName;
+                    user.Email = customer.Email;
+                    user.FirstName = customer.FirstName;
+                    user.LastName = customer.LastName;
+                    user.CreatedOn = DateTime.Now;
+                    user.IsActive = true;
+                    customerDAO.CreateCustomer(user, conn);
+
+                    message = "Customer Account Created Successfully.Your Customer Id is " + cusId;
+                }
                 
-                Guid cusId = Guid.NewGuid();
-
-                Customer user = new Customer();
-                user.UserId = cusId;
-                user.UserName = customer.UserName;
-                user.Email = customer.Email;
-                user.FirstName = customer.FirstName;    
-                user.LastName = customer.LastName;
-                user.CreatedOn = DateTime.Now;
-                user.IsActive = true;
-                customerDAO.CreateCustomer(user,conn);
-
-                message = "Customer Account Created Successfully.Your Customer Id is "+ cusId;
                 return message;
             }
             catch (Exception exp)
             {
                 conn.Rollback();
                 return exp.ToString();
-                throw;
             }
             finally
             {
@@ -70,15 +102,23 @@ namespace StoreAPICore.Controller
             {
                 conn = new DBConnection();
 
-                customerDAO.DeleteCustomer(customerId,conn);
-                message = "Customer Account deleted Successfully.";
+                if (customerDAO.CustomerExists(customerId,conn))
+                {
+                    customerDAO.DeleteCustomer(customerId, conn);
+                    message = "Customer Account deleted Successfully.";
+                }
+                else
+                {
+                    message = "Customer Account does not exists...";
+                }
+
+                
                 return message;
             }
             catch (Exception exp)
             {
                 conn.Rollback();
                 return exp.ToString();
-                throw;
             }
             finally
             {
@@ -89,22 +129,28 @@ namespace StoreAPICore.Controller
             }
         }
 
-        public List<Customer> GetAllCustomers()
+        public DataTable GetAllCustomers()
         {
 
             List<Customer> customerList = new List<Customer>();
-
+            DataTable dataTable = new DataTable();
+            
             try
             {
+                dataTable.Columns.Add("Data", typeof(List<Customer>));
+                dataTable.Columns.Add("Message",typeof(string));
+
                 conn = new DBConnection();
                 customerList = customerDAO.GetAllCustomers(conn);
-                return customerList;
+                
+                dataTable.Rows.Add(customerList,"Data Successfully retrieved..");
+                return dataTable;
             }
             catch (Exception exp)
             {
                 conn.Rollback();
+                dataTable.Rows.Add(null,exp.ToString());
                 return null;
-                throw;
             }
             finally
             {
@@ -121,15 +167,23 @@ namespace StoreAPICore.Controller
             try
             {
                 conn = new DBConnection();
-                customerDAO.UpdateCustomer(customer,conn);
-                message = "Customer details successfully updated.";
+
+                if (customerDAO.CustomerExists(customer.UserId, conn))
+                {
+                    customerDAO.UpdateCustomer(customer, conn);
+                    message = "Customer Account updated Successfully.";
+                }
+                else
+                {
+                    message = "Customer Account does not exists...";
+                }
+                
                 return message;
             }
             catch (Exception exp)
             {
                 conn.Rollback();
                 return exp.ToString();
-                throw;
             }
             finally
             {
